+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Bug Fixes
 * Fix ring finger to foot collider having half the size from VRChat update 2023.1.2.
 * Fix crash when a material swap gets merged into a blob of meshes and its mesh index is non-zero. [(more)](https://github.com/d4rkc0d3r/d4rkAvatarOptimizer/issues/28)
+* Fix always converting vertex colors to full float32 instead of using unorm8 if it was already unorm8.
 
 ## v1.10
 ### Features